Add files via upload
Let's begin with terms:

Things with use-values are stuff that you can pick up and equip or use to get a bonus or benefit. In real life the equivalent is tools, clothing, food, and other things that you purchase to use.

Commodities are things players wants. This includes useful gear or unlocks that are purely aesthetic and convey no useful benefits. Commodities are anything people desire to have and can gain. The cost of the commodity is not tied to their use-value. A luxury item might cost real world money but convey no in-game benefit except be aesthetically pleasing, or others may be very useful and easily attainable for low-in game currency.

When we extend this to the material world, we see the same thing. Everything we purchase is a commodity, but their material value is divorced from their social value and use value. A pair of luxury shoes might be sold for hundreds of thousands of dollars and take the same time & materials to produce as an off-brand knockoff. Yet the branded version is valued and priced higher.

Games themselves too are commodities. The video game industry is worth billions of dollars now, and is still growing fast, with most household having at least one gamer. Yes, that mostly includes mobile games, but the executives at those companies doesn't care about your artistic snobbery, just profit.

As we'll get more into later, the price of these games as a commodity that players pay is pretty arbitrary. A game that you could potentially spend thousands of hours in could cost $15, or $60 (or even more if you have to download a game while living in australia), and short games could be anywhere from completely free to $60. An older game might be sold with all DLC for $15 now, or be arbitrarily set at $30 depending on the publisher. The price of the commodity is clearly not set by the potential use-value the players get out of the games.

If you play online games like MMORPGS, you know that the value of your account, were you to break TOS and sell it, would be proportional to the amount of work (labour) that you put into your account. An account that is developed for years with lots of unlocks and in-game currency and maxed level characters is going to be more valuable an account with only 1 maxed level character, and a completely new account will have no value.

All games were created from the labor of the developers (and the many supportive roles that don't directly create code), just as all labor in real life creates all commodities. however, the amount of labor per piece of work is difficult to determine on the outside. You can't tell what parts of the game were taxing and time consuming to create, and which were easy things that were relatively fast to do. It can also be difficult on the surface to tell which games had more developer labor time invested in them versus less. It's only after investigating do you realize how long games take to develop, and the ways game companies cheat long game dev times by re-using engines and assets, and that experience of the developers, code language & organization of the project contribute quite a bit to either increasing or decreasing the required labor time of the final commodity.

In real life, one simply has to think about the jobs & hobbies of people they know or watch an episode of How It's Made to see how most things that look easy in terms of labor time are actually quite complicated or require time to develop the skills to do properly. In this too, the labor time in these items is usually completely hidden from us. And therefore, what becomes valuable to us is not necessarily things that require lots of time and effort to produce. The pricetag of what we buy is instead set at what marketers think people will pay, and they try to increase that price from the labor & material value as much as possible.

And yet the workers receive only a fraction of the profit from the labor THEY produced. In the games industry, game developers are suffering from rampant labor abuses such as wage theft from unpaid overtime and being pressured to work long hours, while the CEOs take in multimillions every single year. (and of course, a much better benefits package)

Let's use EA games as an example. The EA CEO is expected to make around $40 million dollars in this year alone. The CFO made another 15 million. Split between the full 9,800 employees of EA, that would be 6,632/year to every single employee from part time tech support, to IT, to accounting. This doesn't seem like a lot on it's own, but consider that federal minimum wage at full time is only $13,920, so that's potentially an extra 30% a year to the lowest wage workers just from redistributing just two people's paychecks.  Secondly, even just that small amount is double what most Americans got as a stimulus check during an mass pandemic. And finally, this is the amount that is being skimmed off every single worker's paycheck every single year. 6,632/year is higher than the amount a worker might be paying in taxes to the state and federal government, but in return nothing is provided except the workplace building and workplace equipment. 

Capitalism gets us coming and going. The worker is forced to sell their labor to sustain themselves, but they don't even get a fair deal out of it. At every turn, there is exploitation. A person is first transformed from a thinking breathing human into a 1-dimensional worker that is only allowed to perform set tasks and must suppress their natural human needs and desires, like wanting to be around their family or wanting to be outside on beautiful days.

After the Worker is molded, they are exploited further by not being paid fairly for their work, such as working 70 hrs a week but only paid for 40. The next step of exploitation comes at profit time, when the executives have the choice to take the billion dollar profit (created 100% from the workers) from that year and choose to distribute it unequally and secretively, while creating a culture of silence around organizing and pay. These executives will always choose to increase shareholder profits and bonuses to themselves rather than distribute it fairly. In most cases, they are literally legally required to seek and return as much shareholder profit as possible. The system forces them to exploit their workers. The same thing is almost certainly happening at your job, too. The numbers may be different, but the same mechanism of exploitation is there.